flib: migrate MC flib to version 0.6
There are following major changes in MC flib 0.6:
- Now the MC returns portal offsets instead of portal physical addresses
- Now the cfg struct used by dprc_create_container() has a new field 'label'
  that needs to be initialized. For now, we initialize it as null.

Reviewed-on: http://git.am.freescale.net:8181/34058
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>

+++ dpmac_commands.c
@@ -245,8 +245,8 @@ static int print_dpmac_attr(uint32_t dpmac_id,
 		(target_obj_desc->state & DPRC_OBJ_STATE_PLUGGED) ? "" : "un");
 	print_dpmac_link_type(dpmac_attr.link_type);
 	print_dpmac_eth_if(dpmac_attr.eth_if);
-	printf("maximum supported rate %llu Mbps\n",
-			(unsigned long long)dpmac_attr.max_rate);
+	printf("maximum supported rate %lu Mbps\n",
+			(unsigned long)dpmac_attr.max_rate);
 
 	error = 0;
